+++ include/cowl_nary_type.h
@@ -15,7 +15,7 @@
 
 COWL_BEGIN_DECLS
 
-/// Represents the type of @type{CowlNAryBool} and @type{CowlNAryData}.
+/// Represents the type of @type{CowlNAryCls} and @type{CowlNAryData}.
 typedef enum CowlNAryType {
 
     /// @name Types